[JS] change perf config
reducing the number of testRuns to 30 * numRuns helps speed up the test duration
while it won't harm the measurements because the results are somewhat consistent
after the warmup. A warmup is introduced in the next patch.

The other numeric changes are done, because echoComplexStruct and echoByteArray
are much slower in comparison to the other tests and thus less runs make sure
that all tests take about the same time.

Change-Id: I70a20ac29ecf792dd0b7ec8f147fba1ef9e49741