[JS] added Promise.reject again to joynr.load
Promise.reject is necessary because return will enter
the then case and not the catch. Throw would enter the catch
case too and therefore would almost be the same but is
theoretically an api change since if joynr.load would not be
catched it would still throw while reject would no nothing.

Change-Id: I8fb85b2b8cd4a0c9735636a530c0056cbc91b989