[JS] Performance-tests: add warmup
Performance optimizations by the v8 engine are done lazily on hot code
which was run at least twice and which has enough type information
available. In order to reach peak performance it is necessary to
wait till the v8 optimizations are done. The warmup period takes care
of that. The performance during the warmup period was on average
about 20% slower. It is assumed that the warmup measurement duration
is a lot longer and therefore performance at the start is actually
even slower.

Change-Id: Ibc1063856943d82c3bb1036c7e99bb589e302b8e