[JS] perf ProxyAttribute remove encapsulation for performance
This patch removes the encapsulation with Util.forwardPrototype.
The result is that functions are no longer bound and the dummy
Interface-Object is no longer created. This should improve
performance.
But the price for removing the encapsulation is that the functions
become context sensitive and can no longer be used for a callback.

Instead of binding the functions to the correct context, currying
was used to create a new function each time, which can be used
as a callback once again. This costs a bit more memory, but the
speed improvement compared to the bind is comparably high and since
it's only one function per Attribute, the cost is not too high.

Change-Id: Idd0ac791d95f36453149544608529edc1718c11e