[JS] performance tests: use different config files
Earlier on all arguments were passed via command line to the
performance tests. This had to advantage of not having to change
code when parameters were changed. But this was also very
complicated as more and more parameters were needed. Having a
config file fixes that. But on the other hand the ability to
easily change parameters was lost, because code had to be changed
and this is bothersome when switching commits.
This commit uses the best solution of both worlds. Settings can be
preconfigured in config files and with the command line it can be
configured which config file is to be used.

Change-Id: I38377700a99401713fd6f2f38a7651478a4ccc74