[JS] optimize function calls in MessageRouter
messageRouter.route used to create a new function Object forwardToRouteInternal
each time it was called. This was avoided by two actions.
1)Array.prototype.map takes a context argument which is being used to pass
the joynrMessage argument to a changed forwardToRouteInternal function
outside of MessageRouter.route.
2) instead of using resolveNextHopInternal(..).then(forwardToRouteInternal)
a second function resolveNextHopAndRoute was created which has the same
functionality as the call above. In order to avoid code duplication
findAddressFromPersistency(participantId) was extracted, which was overridden
in case of persistency being undefined and thus saving an if Statement call.

Change-Id: I73beae7ccd1b71920205cecfea41d1131d4de543