[Java] Change log level
The information about successfully processed MQTT request is more
important than the other case where no message id was removed from the
queue, as the processed message might not even be a MQTT message. The
message router triggers the messageProcessed callback regardless of the
type of message it processed or its origin.

Change-Id: I64462be04ccd25cc497825df3599ebe31afbcc3e