[Java] Reuse global discovery proxy for capabilities freshness update
The proxy for the capabilities freshness update (touch method
of the global discovery interface) was recreated for every
call (every PROPERTY_CAPABILITIES_FRESHNESS_UPDATE_INTERVAL_MS
milliseconds).
This created unnecessary cpu load from the proxy building
process and unnecessary memory usage because of routing entries.

Once built, the proxy can be reused for subsequent calls because
the proxy parameters are always the same.

Change-Id: Ie1872b2acdd84d0313c8bbea060f4ae3406f2d1a