[Java] Do not call setParentRouter before the proxy is ready
MessageRouter.setParentRouter is called in the joynr runtime's
constructor.
If setParentRouter is called before the proxy is ready (i.e.
before the corresponding connector has been created),
setParentRouter blocks the calling thread until the connector
is set in the proxy or the proxy creation times out (after
discoveryTimeout).
Since setParentRouter is called by the ArbitratorRunnable thread
in ConnectorFactory.create during the creation of the connector,
this results in a deadlock (until discoveryTimeout is reached)
and a failing runtime creation.

Change-Id: I6515e222f85a163e4adcafa51ecfa07d852c1bec