[JS] remove __proto__ manipulation
There is an eslint rule which forbids __proto__ manipulation,
because it is deprecated since es 3.1
It was only used to trick jslint.
Object.setPrototypeOf still does the same, but is not deprecated.
Generally speaking removing Object.setPrototypeOf should still
be considered in the future since it is bad for the performance.

Change-Id: I8284c56de6c248121e2ff8db4e8d870202a1b33e