[JS] set Util.ensureTypedValues to Typing.augmentTypes
when joynr is loaded Util is placed on joynr.util.Util
and therefore Util.ensureTypedValues is part of the
public api. Typing augmentTypes does exactly what Util.
ensureTypedValues did earlier, except that the calls
are different. Therefore it makes sense to override this
part of the api, in case somebody would use it.

Change-Id: I1db509209905373570937f189f190258dec8c602