[JS] Save NOTICE file for libjoynr-js to src/main/js/META-INF
The NOTICE file 'javascript/libjoynr-js/NOTICE' generated/updated
and checked by the maven-notice-plugin was not used. It was not
included in the joynr node module.
Instead of the automatically checked file, a copy of the NOTICE
file 'javascript/libjoynr-js/src/main/js/META-INF/NOTICE' was
used.
Now the NOTICE file included in the joynr node module is
updted by the maven-notice-plugin directly.

Change-Id: Id470a896a06dd668b49864ac74890af95f602abd