[JS] avoid new when creating DispatchingTypes
When calling new on a function the function is invoked as a Constructor
and memory is allocated to create a new Object which can be addressed
as this within the constructor. If the function returns this or no Object
all it the newly created Object is being returned and can be used in the
rest of the code.

Before this patch most Constructors of DispatcherTypes did not return this,
but rather a manipulated settings Object. Therefore the new Object created by
the constructor has zero references and is ready to be garbage collected.

Having less Objects to be garbage collected is always benefical, but the
effect on the performance is not really measurable.

Simply removing the new in front of the constructor is bad codestyle
critisized by eslint and therefore the constructor functions were moved
to <DispatcherType>.create

Change-Id: I8f38a6b1ba6bf8d3341f2192ca928b1c4eab84f6