[Java] set global address as reply-to for stateless async msgs
Because replies to stateless async requests should be able to
be routed to any participant in the application cluster from
which the request originated, the reply-to address for
stateless async messages is set to the global address, which
means that messages directed at it can be picked up by any
of the cluster nodes.
Normally, a node-specific reply-to address is used, to ensure
that replies for a 'normal' synchronous request are routed
to the correct node, on which the request-specific callback
instance is held in memory.

Change-Id: I48d24b2ef543f11cdc6ade321b602c18072c8516