[Java] use separate manager for request reply IDs
In order to be able to customise how stateless async callback IDs are
obtained, separate out the logic for calculating them from a reply
into its own class.
This way we can then provide a cluster-specific implementation which
stores the stateless callback ID by the requestReplyId in a distributed
store (e.g. Redis) and when a reply is received, we can query that
from any node in the cluster using the request reply ID from the
Reply object in order to ascertain which stateless async callback ID
should be used.

Change-Id: Ia2d71c7f30f5f3761205dc24f6e17b096b29c228