[Java] Add major version of Franca interface to ParticipantIdStorage key
Allow registration of multiple version of a provider interface on
the same domain via the same runtime.

This was not possible before because the participantId of the
provider registered first was reused for subsequent
registerProvider calls for the same interface and domain.
The last provider instance replaced all other version of that
provider interface (RoutingEntry and DiscoveryEntry).

old key format:
<PREFIX><DOMAIN>.<INTERFACE_NAME>
new key format:
<PREFIX><DOMAIN>.<INTERFACE_NAME>.v<MAJOR_VERSION>

Change-Id: I974210538b92178230711b3df88e02f40a454ac3