[Java] use type 3 UUID for stateless participant ID
Instead of sending the full stateless participant ID over the wire, create
a type 3 hashed UUID thereof, keep a directory mapping the UUIDs to the
full form, and send the UUID over the wire and reverse lookup the full
form for finding the relevant stateless async callback for replies.

We also optimise the IDs being used so that less string manipulation is
necessary in order to perform the routing of the replies.

Change-Id: Ib633b047df4d5d698532fd1b867155c8470869d8