[JS] rewrite WebSocketLibjoynrRuntime using ES6 Class
This change is needed to use ES6 class mechanics allowing
the usage of the `extends` keyword to use inheritance in
the following patches.

The major difference is that earlier on variables and
methods were defined in the constructor which made it
so that the scope of the methods allowed the variables
to have infinite lifetype.

The variables shared between functions were kept as
member variables of the runtime while the other ones
can be garbage collected.

By saving more variables on the `this` context they
are potentially accessible from the outside
but this would be invalid usage.

Change-Id: I6fb38252427e8fa3b0bd02485e9f2003b1f4fc7a