[JS] extract common functionality into JoynrRuntime
The extracting of the members and the shutdown method is
rather straight forward itself.

But for start it was more difficult as there are some checks
at the beginning which can be extracted and others in the middle.

Therefore initial checks whether the provisioning is correct
were extracted into super.start and most of the logic got
extracted into _initializeComponents which has to be called
with a few preparatory arguments.

Change-Id: I2ea9928a3192562b11982b2be1da3a090e46336d