[Java] Integrate message persistence into MessageQueue
Implements an optional message persistence hook-in so that users of
joynr can provide a mechanism to externally persist joynr messages.
This is useful for the case where the runtime terminates abnormally and
certain messages should not be lost. When incoming message is put into
the queue, the persistence implementation (if present) is given the
chance to persist this message. The decisions on whether and how to
persist the message are taken by the persister implementation.

At startup the persistence implementation is asked for all messages it
has persisted for this specific message queue and these are added back
into the in-memory queue for processing. Once a message has been
successfully passed to joynr for processing, the persistence
implementation is asked to remove it from its persistence.

Change-Id: I90fa9a1293a36d32f10dfe69b733742413e6487a