[SIT] Log errors in run-onboard-sit.sh as "SIT RESULT error"
Errors should be logged with the key words "SIT RESULT error"
because run_sit.sh searches for "SIT RESULT" lines in the logs
to count the successfull test cases and to detect errors.
Every result line which does not contain "SIT RESULT success"
is treated as error. Without the keywords "SIT RESULT" these
errors lead to false positive test executions because the error
is not detected by the main run script.

Change-Id: I067e7d7fd2cad9483bad57e6ac671385b2f234aa