[Java,Test] Do not reuse MessageQueue for multiple MessageRouter instances
If the same instance of MessageQueue is used in multiple parallel
instances of MessageRouter, there is a race condition between the
MessageWorker threads of the MessageRouter instances.
The default instance of MessageRouter created in @Before had to
be shut down in the test cases which need another instance with
different settings.
This is now avoided by removing the binding for MessageQueue from
the Guice module which is reused in some test cases to create a
MessageRouter with different configuration. Without explicit
binding, a new instance of MessageQueue is created by Guice.

Change-Id: I93c59fc71a21ad1b8f7ef8e48077b6e2f8e44a7e