[Java] Use shorter String representation for UUIDs
Using base64 encoding the length of the resulting String drops to 22
characters. In general, there are multiple options how to represent the
128bits of a UUID as a String:

a) 4bits for one character: This is the default UUID string
representation. 128bits are converted to 32 chars in the range '0' to
'f' (i.e. hexadecimal numbers). Additional four chars '-' are placed
inbetween resulting in total length of 36.

b) 6bits for one character: This is what base64 does and the method we
choose here. The 128bits are represented by 22 chars in the range
[a-z, A-Z, 0-9, -, _]. We use the URL-friendly alphabet in order to
avoid the '/' and '+'  character that are included in the basic base64
alphabet but are special characters for mqtt topics.

c) 8bits for one character: This would correspond to ASCII and will
result in a string of length 16. However, in the ASCII table 0-127 some
characters are not displayable and therefore we can not use this compact
representation.