[Generator,Java] Fix generation of Deferreds for provider interface
Deferreds (part of the return value of the generated provider
interface) were not generated correctly:
The deferreds have to be generated depending on the out paramters
(resolve method for the out parameters) and the errors (reject
method for the error enumeration) defined in the interface but
the error enumeration was not taken into account.

Example:
2 methods with the same output parameters, one method with
errors, one method without errors:
Only one deferred was generated and used for both methods, either
the deferred without the errors defined for one method or the
deferred with the errors.
The problem is similar for 2 methods with the same out parameters
and different errors.