[Java] annotate JoynrRuntimeException with ApplicationException
In order for the EJB runtime not to wrap the joynr runtime exceptions
in EJB rollback exceptions the JoynrRuntimeException is now
annotated with ApplicationException as per EJB spec.

The JEE API is included as provided so that it will compile. Due to the way
annotations are implemented, it is ignored in runtimes where the dependency
does not exist.

Change-Id: Ic0380010ecbbe39f2e4628efb7a465df5a81414b