[Java] Add new public methods to the GCDClient
The new methods include gbid as a parameter. This way the GCD proxy
calls can be augmented (via custom header) with the information which
GCD (i.e. in which GBID) is the target of the proxy call. By now the new
methods are empty. The old methods call the new ones with the default
gbid as a parameter. The old touch method calls the new touch method for
every known gbid.