[Java] Move functionality from ProxyInvocationHandlerImpl to abstract class ProxyInvocationHandler
such that the parent abstract class (ProxyInvocationHandler)  has no dependency
on routing layer, and only subclasses can introduce dependencies on
MessageRouter. Subclasses may need functionality that was present in
ProxyInvocationHandlerImpl, which has a MessageRouter dependency. To avoid
duplicate code, or introducing another abstract class in between
ProxyInvocationHandler and ProxyInvocationHandlerImpl, the common functionality
is moved to the parent class.