[Android] - Android slf4j fix for simple logging
In order to configure the internal loggers created by joynr runtime
it's necessary to save the logging level configured by the developer.
The previous implementation was using the log level configured on any
already available loggers but in case there were none loggers configured
the loglevel value was not being saved and later reused in new loggers
created by the classes inside the joynr runtime.