Improve Gradle builds for Android projects
Some Android projects are relying on a copy task that moves the needed
compiled .jars to the appropriate output folder. However, since this
task isn't specifying outputs, Gradle can't correctly optimize the
process and may sometimes fail to understand that the task needs to be
re-run. Therefore, tell Gradle how to determine whether the task is
up-to-date or not.