[Java] Fix types of array responses in StatelessAsyncReplyCaller
When calling a method that returns an array on a stateless async proxy,
the StatelessAsyncReplyCaller throws:

IllegalArgumentException: argument type mismatch

Since arrays itself do not have a _typeName field, they are deserialized
as Object arrays, even though the included elements are deserialized as
expected to their proper types. Therefore the RpcAsyncRequestReplyCaller
converts the responses to their proper types:
https://github.com/bmwcarit/joynr/blob/624bfa5c63512cebe7e7b66a7d8ca9752372cb27/java/core/libjoynr/src/main/java/io/joynr/dispatching/rpc/RpcAsyncRequestReplyCaller.java#L79

However, this functionality was missing in the StatelessAsyncReplyCaller.