[Java] Ignore local LCD store in global only lookup for cacheMaxAge = 0
This is required for lookup by domain and interface to be able to
force a remote (global) lookup.
It makes no sense for lookup by participantId because
participantIds must be unique, i.e. a remote entry is the same as
the local entry. If there is a duplicate entry (because of
invalid configuration), the remote entry is ignored by the LCD.