[Java] Fix routing entry refcount handling for subscription replies
Only reduce refcount of the proxy's routing entry (recipient) if
the provider (sender) has an InProcessAddress.

The refcount must not be decremented if the provider is running
in a libjoynr runtime connected to the CC e.g. via WebSockets.

The refCnt of the proxy routing entry in the CC has only been
increased by the PublicationManager if the provider is running
in the CC runtime (InProcessAddress). Publications from WebSocket
providers cannot be routed in the CC if the refcount has been
decremented on the SubscriptionReply message.