[Java] Store proxy in SubscriptionManager
If the proxy gets out of focus in the application, it will be garbage collected
and its routing entry will be removed by a periodic cleanup job.
Afterwards, a publication cannot be delivered to the listener.

To keep the subscription alive, the proxy instance should be stored together
with the listener in the SubscriptionManager.