[Java] Extend ttl of lookup between Libjoynr and CC by epsilonMs time
The epsilon shall only make sure that the reply from the lookup
call is not lost because of a race condition between the local
timer for the expiration of the proxy call and the delivery of
the result. The lookup call will usually return immediately after
the remaining ttl expired.