[C++] Calculate ttl for lookup in Arbitrator from arbitration deadline
The arbitrator performs the lookup with the ttl set to the
remaining discovery timeout + 10 seconds grace period (epsilon).
The epsilon shall only make sure that the reply from the lookup
call is not lost because of a race condition between the local
timer for the expiration of the proxy call and the delivery of
the result. The lookup call will usually return immediately after
the remaining ttl expired.