[Java] Make multicastListeners unique
Changed the datastructure holding the multicastListeners to be a set
instead of a list to make sure the listeners are unique. This is done
to avoid an issue when PublicationManagerImpl.entryAdded() is called
multiple times which would previously lead to a duplicated entry
in the multicastListeners. This would also happen when the same
provider is registered multiple times.