[Java] Fix dead lock in PublicationManagerImpl.addSubscriptionRequest
If a thread has locked addRemoveLock in PublicationManagerImpl
.addSubscriptionRequest but has not yet executed
providerDirectory.get() inside the synchronized(addRemoveLock)
block, a dead lock could occur if another thread tried to
register or unregister a provider at that time.
Provider registration / deregistration adds / removes an entry in
providerDirectory which locks the providerDirectory and notifies
its listeners. PublicationManagerImpl is such a listener and
requires addRemoveLock which is already locked in
addSubscriptionRequest() which requires the lock in
providerDirectory for providerDirectory.get().