[Java] Fixed dependency problems in joynr-generator-standalone
Investigation revealed that Maven dependency resolution ignored our
specified artifact version provided the dependency section of
another artifact required the same artifact as transitive dependency
with a range rule where our version did not match. In that case Maven
used the latest available version within that range. In this specific
case we ended up with a version of some artifact which was no longer
compatible with Java 8 finally causing the build to fail.

The version settings have now been moved from dependency section into
a dependencyManagement section to take effect.

Furthermore the maven-enforcer-plugin has been introduced with
dependencyConvergence and banTransitiveDependencies checks. This
reports versions that differ from each other for the same dependency
artifact and transitive dependencies that have not been pinned yet.

In order to cover these, extra pinning has been added as well as
extra ignore rules for the maven-dependency-plugin which is unable
to detect some artifacts are used.