[Java] Fixed stateless asynchronous remote procedure calls
Only a single stateless async reply could be handled in CC runtimes
since the improved handling of routing entries to reduce the memory
usage was introduced:

The routing entry of the stateless async CallbackHandler /
StatelessAsyncCallback was decremented and potentially removed after
processing the message.

Further reply messages could not be routed anymore and were dropped.