[C++] Compatibility with CMake 3.13
With CMake 3.13, compilation of the cluster controller fails because
libJoynr and libJoynrClusterController are not linked to the mosquitto
and MoCOCrW libraries. This is most likely a deficiency in CMake, as it
works with a newer version. Extend the use of the work-around we use for
earlier versions also for CMake 3.13.