[Java] Return empty GBID in results when lookup was called with empty GBID
In case of bad provisioning, the only known GBID is the empty GBID. On a
successful lookup, if the received address contained a non-empty GBID,
communication with the discovered provider would not be possible since
the GBID is not known.

Now, the received GBID is replaced by an empty GBID if the lookup was
initiated with an empty GBID as well.