[TS generator] added 'default' provider implementation creator to generated ts code
The generated provider implementation can be used for testing.
It implements every function (methods as well as attribute
getters and setters) to return either void, or an empty object
depending on the specification. The return values are set as
attributes <functionName>ReturnValue in the created
implementation. This allows overwriting the return values (if not
void) to change the behavior of the provider.
NOTE:
The generated code does not fully support interfaces with
overloaded methods. It is not possible to define different return
values, even if the return types in the interface are different.
All overloads of a method will return the same value.