[Java] Do not duplicate multicasts for receivers in the same runtime
Messages in CcMessageRouter were scheduled per recipient.
This resulted in multiplication of multicast messages if multiple
multicast receivers were located in the same runtime (with the
same address, e.g. WebSocketClientAddress, BinderAddress).
Because of this message multiplication, multicast receivers in
such runtimes received the same multicast publication multiple
times.

To prevent multiplcation of messages, only a single message is
scheduled now for multiple recipients with the same address.
Instead of scheduling messages only for a single participantId,
each message now contains a set of recipients with the same
address. This ensures that the message will also be delivered if
one of the recipients got removed from the routing table.