[C++] Fix lock in TaskSequencer potentially causing a deadlock
The lock in the TaskSequencer is intended to protect the internal task
queue from simultaneous read/write operations. Therefore it is not
necessary for a thread to hold this lock while executing a task that
is already removed from the queue. This was the case, and could lead
to deadlocks if a second thread was also holding a lock and trying to
manipulate the task queue, while the first thread was waiting for the lock
held by the second thread.