[Java] Prepare Mqtt Clients for external connect and disconnect signals
This allows the HivemqMqttClientFactory to trigger the start and stop
of Mqtt Clients after receiving external signals. It also allows the
logic behind starting the clients to be dependent on the external
signals of what the MQTT connection status should be.

To be able to add tests to the start and stop logic, it was needed to
create a separate class HivemqMqttClientCreator separate from the
HivemqMqttClientFactory which deals with all the Hivemq Mqtt Client
creation logic that was already present in the factory, making the
factory class look cleaner, and more testable, and the overall
code more organized. This class is then injected into the factory
using Guice.