[C++] Fix race condition w.r.t. awaitGlobalRegistration
A race condition occured in LocalCapabilitiesDirectory when an newly added
entry by 'add' with awaitGlobalRegistration == false was removed by previously
scheduled and delayed 'remove' job executed from TaskSequencer.

The value of awaitGlobalRegistration is now saved and associated with
participantId so that when 'remove' is called, the behavior depends on
the stored value.

In case where awaitGlobalRegistration == true, the local entry removal process
happens when the job is executed from TaskSequencer.
If awaitGlobalRegistration == false, the local entry is removed before the
'remove' job is scheduled and no entry is deleted in completion callbacks.