[Jee] Improve exception handling in radio-jee
Method addRadioStation may throw joynr.exceptions.ApplicationException
with DUPLICATE_RADIOSTATION error.

Previously radio-jee-customer would catch this exception and rethrow it as
WebApplicationException. This, however, did not preserve the error message
of actual exception.

In order to fix this, we have implemented custom exception class and
ExceptionMapper dedicated to it. This way we can throw proper exception
and preserve its original error message, and also set correct status code,
and this exception will be a properly formatted response.

Additionally, custom exception class takes message as an argument, so it
does not rely on specific exception type only, making it ready to use
with other exceptions with no modifications whatsoever.